Moved RBAC functionality to new class
