+++ client/src/services/Constants.js
@@ -14,3 +14,4 @@ export const AUTH_SUCCESS_MESSAGE = "Success! Reload the page to load your new c
 export const AUTH_CONFLICT_MESSAGE = "Role already exists. Restart the app to reload your credentials";
 export const AUTH_FORBIDDEN_MESSAGE = "Forbidden. You do not have permission to changes roles on this twin instance";
 export const AUTH_NOT_FOUND_MESSAGE = "Twin not found";
+export const TWIN_DATA_OWNER_RBAC_ID = "bcd981a7-7f74-457b-83e1-cceb9e632ffe";
